move io_utils into tests/
