+++ src/tests/brain/timer_test.py
@@ -1,7 +1,7 @@
 import unittest
 
 from vex import Brain, SECONDS
-from ...testing.io_utils import replace_stdin
+from ..io_utils import replace_stdin
 
 
 class TestBrainTimer(unittest.TestCase):
