+++ src/tests/drivetrain_test.py
@@ -2,7 +2,7 @@ import unittest
 
 from vex import DriveTrain
 from vex import Motor, Ports, DEGREES, PERCENT, FORWARD, REVERSE, MM, LEFT, CurrentUnits
-from ..testing.io_utils import replace_stdin
+from .io_utils import replace_stdin
 
 
 class TestDriveTrain(unittest.TestCase):
